Add user's legal name in registration process move to user fragment (unfinished)
